feat(davinci): split DOM patch facts
