+++ crates/vize_s1_to_s2/src/emit.rs
@@ -86,6 +86,7 @@ mod once;
 mod options;
 mod outlet;
 mod outlet_props;
+mod patch;
 mod prefix;
 mod props;
 mod props_bind;
