+++ crates/vize_s1_to_s2/src/emit/props.rs
@@ -4,19 +4,16 @@ mod constness;
 mod static_expr;
 mod ts_view;
 
-use alloc::vec::Vec as StdVec;
-
-use vize_s0::String;
-use vize_s2::op::{Attribute, BindingOp, OnOp};
+use vize_s2::op::{Attribute, BindingOp};
 
 use super::EmitCx;
 use super::EmitError;
 use super::UnsupportedReason as Reason;
 use super::buf::Buf;
-use super::on::{admit_on, event_key_for, needs_hydration};
+use super::on::admit_on;
 pub(in crate::emit) use constness::handler_is_constant;
 pub(super) use constness::{bind_value_is_static_patchless, bind_value_text};
-use static_expr::bind_value_uses_legacy_patchless_runtime_expr;
+pub(super) use static_expr::bind_value_uses_legacy_patchless_runtime_expr;
 
 pub(super) use super::props_bind::{
     BindName, StaticBindKeyCasing, bind_name, emit_dynamic_bind_pair, has_prop_modifier,
@@ -24,11 +21,6 @@ pub(super) use super::props_bind::{
 };
 pub(super) use super::props_object::{Piece, PropsObjectOptions, emit_props_object, pieces};
 pub(super) use super::props_value::bind_value;
-
-pub(super) struct Patch {
-    pub flag: i32,
-    pub dynamic_props: StdVec<String>,
-}
 
 #[derive(Clone, Copy, Default)]
 pub(super) struct BindPropsOptions<'a> {
@@ -103,160 +95,6 @@ fn admit_bindings_inner(bindings: &[BindingOp<'_>], allow_once: bool) -> Result<
     Ok(())
 }
 
-#[allow(clippy::too_many_arguments)]
-pub(super) fn bind_patch<'a>(
-    bindings: &[BindingOp<'a>],
-    is_component: bool,
-    if_key: Option<&str>,
-    for_item: bool,
-    is_ts: bool,
-    constant_handler: &dyn Fn(&str) -> bool,
-    handler_is_cached: &dyn Fn(&OnOp<'a>) -> bool,
-    caches_handlers: bool,
-) -> Patch {
-    if super::merge::has_object_spread(bindings) {
-        return super::merge::object_patch(
-            bindings,
-            is_component,
-            if_key,
-            for_item,
-            is_ts,
-            constant_handler,
-            handler_is_cached,
-            caches_handlers,
-        );
-    }
-    let mut flag = 0i32;
-    let mut dynamic_props = StdVec::new();
-    for binding in bindings.iter() {
-        match binding {
-            BindingOp::Bind(bind) => match bind_name(bind) {
-                _ if is_emitted_key_bind(bind, if_key) => {
-                    if for_item && is_dynamic_bind_name(bind) {
-                        flag |= 16;
-                    }
-                }
-                Ok(BindName::Static(raw_name)) => match raw_name {
-                    "ref" => flag |= 512,
-                    "class"
-                        if bind_value_is_static_patchless(bind, is_ts)
-                            || (!for_item
-                                && bind_value_uses_legacy_patchless_runtime_expr(bind)) => {}
-                    "class" if !is_component => flag |= 2,
-                    "style"
-                        if bind_value_is_static_patchless(bind, is_ts)
-                            || (!for_item
-                                && bind_value_uses_legacy_patchless_runtime_expr(bind)) => {}
-                    "style" if !is_component => flag |= 4,
-                    "key" => {}
-                    key if key.ends_with("Modifiers")
-                        || bind_value_is_static_patchless(bind, is_ts) => {}
-                    _ => {
-                        flag |= 8;
-                        let Ok(key) = static_bind_key(bind, StaticBindKeyCasing::Preserve) else {
-                            continue;
-                        };
-                        let owned = String::from(key.as_str());
-                        if !dynamic_props.contains(&owned) {
-                            dynamic_props.push(owned);
-                        }
-                        if has_prop_modifier(bind) {
-                            flag |= 32;
-                        }
-                    }
-                },
-                Ok(BindName::Dynamic(_)) => {
-                    flag |= 16;
-                    if has_prop_modifier(bind) {
-                        flag |= 32;
-                    }
-                }
-                Ok(BindName::Spread) | Err(_) => {}
-            },
-            BindingOp::On(on) => {
-                if super::on_dynamic::is_dynamic_on_name(on) {
-                    flag |= 16;
-                    continue;
-                }
-                let Ok(key) = event_key_for(on, !is_component) else {
-                    continue;
-                };
-                if !handler_is_constant(on, constant_handler) && !handler_is_cached(on) {
-                    flag |= 8;
-                    if !dynamic_props.contains(&key) {
-                        dynamic_props.push(key.clone());
-                    }
-                }
-                if !is_component && needs_hydration(key.as_str(), on) {
-                    flag |= 32;
-                }
-            }
-            BindingOp::Model(model) => {
-                super::model::patch(
-                    model,
-                    is_component,
-                    &mut flag,
-                    &mut dynamic_props,
-                    caches_handlers,
-                );
-            }
-            BindingOp::VueHtml(_) => {
-                flag |= 8;
-                let key = String::from("innerHTML");
-                if !dynamic_props.contains(&key) {
-                    dynamic_props.push(key);
-                }
-            }
-            BindingOp::VueText(_) => {
-                flag |= 8;
-                let key = String::from("textContent");
-                if !dynamic_props.contains(&key) {
-                    dynamic_props.push(key);
-                }
-            }
-            _ => {}
-        }
-    }
-    // The shipped `NEED_PATCH` gate names `v-model` beside `v-show`, the
-    // custom directives and `ref`. Only a *cached* update handler reaches
-    // the difference: without caching the model always sets `PROPS`,
-    // which suppresses `NEED_PATCH` on both sides.
-    let has_model = bindings
-        .iter()
-        .any(|binding| matches!(binding, BindingOp::Model(_)));
-    if (super::directive::has_runtime(bindings) || has_model) && flag & (2 | 4 | 8 | 16) == 0 {
-        flag |= 512;
-    }
-    if flag & 16 != 0 {
-        flag &= !(2 | 4 | 8);
-    }
-    Patch {
-        flag,
-        dynamic_props,
-    }
-}
-
-pub(super) fn prune_legacy_patchless_dynamic_props(
-    bindings: &[BindingOp<'_>],
-    dynamic_props: &mut StdVec<String>,
-) {
-    for binding in bindings.iter() {
-        let BindingOp::Bind(bind) = binding else {
-            continue;
-        };
-        if has_prop_modifier(bind) || !bind_value_uses_legacy_patchless_runtime_expr(bind) {
-            continue;
-        }
-        let Ok(BindName::Static(_)) = bind_name(bind) else {
-            continue;
-        };
-        let Ok(key) = static_bind_key(bind, StaticBindKeyCasing::Preserve) else {
-            continue;
-        };
-        dynamic_props.retain(|name| name.as_str() != key.as_str());
-    }
-}
-
 pub(super) fn emit_bind_props(
     cx: &mut EmitCx<'_>,
     attributes: &[Attribute<'_>],
@@ -308,18 +146,6 @@ pub(super) fn emit_bind_props(
         cx.buf.push(")");
     }
     Ok(())
-}
-
-pub(super) fn apply_static_ref_patch(attributes: &[Attribute<'_>], flag: &mut i32) {
-    let has_static_ref = attributes.iter().any(|attr| attr.name == "ref");
-    // The shipped gate is the *normal prop* flags alone: NEED_PATCH still
-    // combines with TEXT and NEED_HYDRATION, because neither of those
-    // updates a ref on its own. Only an inlined render function reaches
-    // the difference — elsewhere a handler that sets NEED_HYDRATION is
-    // dynamic and sets PROPS with it.
-    if has_static_ref && *flag & (2 | 4 | 8 | 16) == 0 {
-        *flag |= 512;
-    }
 }
 
 fn has_dynamic_bind_name(bindings: &[BindingOp<'_>], if_key: Option<&str>) -> bool {
