+++ davinci-road/plan/phase-3-records/p3-7.md
@@ -1,0 +1,43 @@
+# P3-7 - VDOM patch facts first slice
+
+This slice starts replacing VDOM patch-flag inference with an explicit DOM
+realization fact surface while keeping shipped DOM output byte-identical.
+
+## Boundary
+
+The slice chooses the S2 to S4 annotation route for now. Current S3 records the
+canonical op graph and static/dynamic partition, but it does not yet carry the
+typed binding payloads that Vue patch flags need: static bind keys, modifier
+realization, cached-handler state, component fallthrough props, and legacy
+patchless runtime expressions. Sending DOM through S3 here would either lose
+those inputs or re-derive them later, which is exactly the P3-7 problem.
+
+This does not close P3-7. The remaining work is to materialize the owner-keyed
+patch facts before VNode emission, connect them to the reactivity-lattice facts
+where binding metadata is available, and re-gate the DOM corpus.
+
+## Decisions
+
+| subject       | decision                                                                                              |
+| ------------- | ----------------------------------------------------------------------------------------------------- |
+| Fact home     | `vize_s1_to_s2::emit::patch` owns `PatchFacts` instead of leaving the decision inside props emission. |
+| Compatibility | Existing `vize_atelier_core::codegen::patch_flag` remains only for compatibility fallback lanes.      |
+| Data carried  | `PatchFacts` carries the numeric flag plus the dynamic-prop key list in source order.                 |
+| Local masks   | `v-for`, `v-once`, `v-memo`, static refs, dynamic slots, and text-child flags stay at VNode position. |
+| Parity rule   | The broad patch-flag fixtures still compare S2 DOM output against the shipped compatibility lane.     |
+
+## Mechanical witnesses
+
+- `cargo check -p vize_s1_to_s2 -p vize_atelier_dom`
+- `cargo test -p vize_atelier_dom --test davinci_s2_patch_flags -- --nocapture`
+- `cargo test -p vize_atelier_dom --test davinci_s2_recent_patch_flags -- --nocapture`
+- `cargo test -p vize_atelier_dom --test davinci_s2_patch_flags_static -- --nocapture`
+- `cargo test -p vize_atelier_dom --test davinci_s2_patch_flags_legacy --features legacy -- --nocapture`
+
+## Follow-up
+
+P3-7 still needs a pre-emission side table that keys `PatchFacts` by owner
+`NodeId`, plus DOM corpus TS-11 and patch-flag equivalence fixtures over the
+materialized table. Once S3 has typed binding payloads, this fact surface can
+move from S2 to the explicit S3 decision stream without changing the VNode
+writer again.
