+++ davinci-road/plan/storage-inventory.tsv
@@ -76,6 +76,7 @@ s1_to_s2	-	crates/vize_s1_to_s2/src/emit/model_key.rs	0	0	1	8	0	0	0	0
 s1_to_s2	-	crates/vize_s1_to_s2/src/emit/on.rs	0	0	1	5	0	0	1	3
 s1_to_s2	emit	crates/vize_s1_to_s2/src/emit/on_dynamic.rs	1	5	0	0	0	0	0	0
 s1_to_s2	emit	crates/vize_s1_to_s2/src/emit/options.rs	1	4	1	8	0	0	0	0
+s1_to_s2	emit	crates/vize_s1_to_s2/src/emit/patch.rs	1	3	1	5	0	0	0	0
 s1_to_s2	-	crates/vize_s1_to_s2/src/emit/prefix.rs	0	0	1	11	0	0	0	0
 s1_to_s2	-	crates/vize_s1_to_s2/src/emit/prefix/aliases.rs	0	0	1	8	0	0	0	0
 s1_to_s2	emit	crates/vize_s1_to_s2/src/emit/prefix/codegen_visitor.rs	1	9	1	11	0	0	0	0
@@ -88,7 +89,6 @@ s1_to_s2	emit	crates/vize_s1_to_s2/src/emit/prefix/slot_defaults.rs	1	7	1	8	0	0	
 s1_to_s2	emit	crates/vize_s1_to_s2/src/emit/prefix/splice.rs	1	4	1	12	0	0	0	0
 s1_to_s2	-	crates/vize_s1_to_s2/src/emit/prefix/strip.rs	0	0	1	7	0	0	0	0
 s1_to_s2	-	crates/vize_s1_to_s2/src/emit/prefix/typescript.rs	0	0	1	11	0	0	0	0
-s1_to_s2	emit	crates/vize_s1_to_s2/src/emit/props.rs	1	3	1	5	0	0	0	0
 s1_to_s2	-	crates/vize_s1_to_s2/src/emit/props/constness.rs	0	0	1	2	0	0	0	0
 s1_to_s2	-	crates/vize_s1_to_s2/src/emit/props/ts_view.rs	0	0	1	3	0	0	0	0
 s1_to_s2	emit	crates/vize_s1_to_s2/src/emit/props_bind.rs	1	5	1	2	0	0	0	0
